consumer/requirements.txt
pyspark==4.0.0
delta-spark
kafka-python
